requirements.txt
altgraph==0.17.3
sounddevice~=0.4.6
librosa~=0.10.1
pyaudio~=0.2.13
pydub~=0.25.1
pillow~=10.0.1
requests~=2.31.0